+++ backstage-port/controller/news.js
@@ -2,6 +2,7 @@ const { Router } = require('express')
 const router = Router()
 const auth = require('./auth')
 const newsModel = require('../database/model/news')
+const swiperModel = require('../database/model/swiper')
 const adminUserModel = require('../database/model/adminUser')
 const categoryModel = require('../database/model/category')
 
@@ -65,7 +66,7 @@ router.get('/getnews',auth,async (req, res, next) => {
     }
 })
 
-// 仅获取新闻Id以及title ,
+// 获取新闻信息
 router.get('/getNewsId',auth ,async (req, res, next) => {
     try{
         let {page = 1, page_size = 10} = req.query
@@ -74,7 +75,8 @@ router.get('/getNewsId',auth ,async (req, res, next) => {
         const dataList  = await newsModel
             .find().skip((page-1)*page_size).limit(page_size)
             .sort({ create_time: 1, _id: -1})
-            .select('_id title')
+            .select('img content title')
+           
             res.json({
                 code: 200,
                 msg: 'success',
@@ -96,6 +98,40 @@ router.get('/getnews/:id',auth,async (req, res, next) => {
     try{
         const {id} = req.params
         const dataList  = await newsModel.findById(id)
+            .populate({
+                path: 'author',
+                select: '_id nickname avatar'
+            })
+            .populate({
+                path: 'type',
+            })
+            // .populate({
+            //     path: 'author',
+            //     select: 'nickname avatar'
+            // })
+            // .populate({
+            //     path: 'type',
+            //     select:'title icon'
+            // })
+
+            res.json({
+                code: 200,
+                msg: 'success',
+                data: dataList
+            })
+    }catch(err){
+         // next(err)
+         res.json({
+            code: 400,
+            msg: '新闻获取失败' + err
+        })
+    }
+})
+// 通过typeID获取新闻内容
+router.get('/getnewsID/:id',auth,async (req, res, next) => {
+    try{
+        const {id} = req.params
+        const dataList  = await newsModel.find({type: id})
             .populate({
                 path: 'author',
                 select: '_id nickname avatar'
@@ -166,12 +202,32 @@ router.patch('/editNews', auth, async(req, res, next) => {
 router.delete('/delete',auth, async(req, res, next) => {
     try{
         let { id } = req.query
-        const deldata = await newsModel.deleteOne({_id: id})
-        res.json({
-            code: 200,
-            msg: '删除成功！！！',
-            data: deldata,
-        })
+        const swiperData = await swiperModel.find({newsId:id})
+        if(swiperData){
+            const deleteswiper = await swiperModel.deleteMany({newsId: id})
+            if(deleteswiper.ok != 0){
+                const deldata = await newsModel.deleteOne({_id: id})
+                res.json({
+                    code: 200,
+                    msg: '级联轮播图及新闻已删除',
+                    data: deldata,
+                    count: deleteswiper.ok
+                })
+            }else{
+                res.json({
+                    code: 200,
+                    msg: '级联轮播图删除失败',
+                    data: deldata,
+                })
+            }
+        }else{
+            const deldata = await newsModel.deleteOne({_id: id})
+            res.json({
+                code: 200,
+                msg: '删除成功！！！',
+                data: deldata,
+            })
+        }
     }catch(err) {
         // next(err)
         res.json({
@@ -181,4 +237,32 @@ router.delete('/delete',auth, async(req, res, next) => {
         })
     }
 })
+// // 删除一条新闻
+// router.delete('/delete',auth, async(req, res, next) => {
+//     try{
+//         let { id } = req.query
+//         const swiperData = await swiperModel.find({newsId:id})
+//         if(swiperData){
+//             res.json({
+//                 code: 400,
+//                 msg: '请先删除此新闻对应轮播图！'
+//             })
+//         }else{
+//             const deldata = await newsModel.deleteOne({_id: id})
+//             res.json({
+//                 code: 200,
+//                 msg: '删除成功！！！',
+//                 data: deldata,
+//             })
+//         }
+//     }catch(err) {
+//         // next(err)
+//         res.json({
+//             code: 400,
+//             msg: '删除失败'+err,
+//             body: req.body
+//         })
+//     }
+// })
+
 module.exports = router
